rename: 페이지 폴더 내부 폴더 구조 변경
