+++ src/routes/Router.tsx
@@ -1,10 +1,10 @@
 import Layout from "@components/layout/common/Layout";
 import SigninPage from "@pages/SigninPage";
 import AuthSuccessRedirectPage from "@pages/AuthSuccessRedirectPage";
-import AllMembersPage from "@pages/member/AllMembersPage";
+import AllMembersPage from "@pages/AllMembersPage";
 import NotFoundErrorPage from "@pages/NotFoundErrorPage";
 import RoutePath from "@routes/routePath";
-import PendingMembersPage from "@pages/member/PendingMembersPage";
+import PendingMembersPage from "@pages/PendingMembersPage";
 import PaymentStatusPage from "@pages/PaymentStatusPage";
 import { RouteObject, createBrowserRouter, Navigate } from "react-router-dom";
